requirements.txt
pyautogui
pillow
python-dateutil
